AVE-1899
Change Confidence to % Observed

Change confidence to percent observed. Right now we are double writing it and that will soon change. I just need to make sure that we double write until everything can be changed over.

+++ Atspm/SpeedManagementImporter/Services/Clearguide/ClearguideFileDownloaderService.cs
@@ -80,7 +80,7 @@ namespace SpeedManagementImporter.Services.Clearguide
                         BinStartTime = entity.BinStartTime,
                         SegmentId = routeEntity.SegmentId,
                         SourceId = sourceId,
-                        PercentObserved = (long)entity.DataQuality * 100,
+                        PercentObserved = entity.DataQuality * 100,
                         Average = entity.Average,
                         Violation = null,
                         ExtremeViolation = null,
